requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
sqlalchemy==2.0.23
alembic==1.13.1
redis==5.0.1
requests==2.31.0
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
aiofiles==23.2.1
Pillow==10.1.0
reportlab==4.0.7
ebooklib==0.18.1
meilisearch==0.28.0
apscheduler==3.10.4
bcrypt==4.1.1
python-jose[cryptography]==3.3.0
jinja2==3.1.2
psycopg[binary]==3.2.3
